Added notes to the #Devel;opment section AND a Gitlab repo deprecation notice.
```
Presently (Jan 2018) config.yml and docker-compose.yml use hard-coded user:password credentials, in real world deployment this should use some abstraction layer.  Examples of secure mechanisms for this would be docker_login or some external call to a AS in your infrastructure ((open)ldap,krb5,saml) that would be defined in an auth or environment variable.

As an example there is a stub file ( `auth_keys_from_cas` ) in this repo that could be used as a file that would have the hashed/encrypted values for tokens/user:pass entries for all auth'd users.

	Example deployment using such a file:
	* 	User authenticates to local domain/AS server
	*	Server passes (over secure channel communication (DTLS or other secure channel within organization's policies) to this file not much unlike a journal write of a auth attempt for admin/sudo rights.
	*	Config.yml calls this file to populate auth'd hosts/users and timeout periods of said tokens (if used), to populate it's auth'd user db or update it on some on-demand or periodic schedule.
```

```
**********************
DEPRECATION of Gitlab repo:

Effective Jan 1, 2018, ALL references to the gitlab will be phased out, as that repo has been deprecated, largely for better workflow and greater contributor intake here on Github.

**********************
```

Signed-off-by: Corey 'linuxmodder' Sheldon <sheldon.corey@openmailbox.org>